Fix player.place event in storage saving
When saving an event into the event history, it will need to extract the
CFTools ID. Before this commit, there were only two fields used for
that, cftools_id and murderer_id. However, some events (e.g.
player.place) use the player_id field for the CFTools ID.

This commit fixes that such events can not be used with cftools-relay.